+++ src/components/pages/course/lesson/VideoLayers/PlaybackSpeedLayer.vue
@@ -14,7 +14,7 @@ export default defineComponent({
   props: {
     show: {
       type: Boolean,
-      required: true,
+      required: false,
     },
     speed: {
       type: Number,
